fix(merge-strategies): fixes array handling for merge and deep
previous implementation allowed some mutations for the returned object to have an effect over
defaults

BREAKING CHANGE: Any usage of the merge or deep functions that relies on the mutation of the
returned object to mutate the defaults argument will cease working. Mutations to the returned object
won't have an effect over defaults.