fix(merge-strategies): fix shallow inconsistent behavior
The behavior of the previous implementation for shallow was inconsistent with the merge and deep
functions, as it didn't use lodash.mergewith under the hood. One of such inconsistencies was the
handling of undefined keys in the data object, which were dealt with using traditional object
assignment.

BREAKING CHANGE: The shallow function will now assign to keys in any inner data object with an
undefined value its value at the defaults object, if any. Previously, the merge and deep functions
did already behave in this manner. As a product of the alignment in behavior with merge and deep,
there might be additional edge cases in which the usage of shallow might break your implementation.
It is suggested that you test your use case after upgrading.